+++ test/plugin/pretty_location_test.rb
@@ -11,7 +11,7 @@ describe "the pretty_location plugin" do
   end
 
   it "uses context to build the directory" do
-    uploaded_file = @uploader.upload(fakeio, record: NameSpaced::OpenStruct.new(id: 123), name: :avatar)
+    uploaded_file = @uploader.upload(fakeio, record: OpenStruct.new(id: 123), name: :avatar)
 
     assert_match %r{^openstruct/123/avatar/[\w-]+$}, uploaded_file.id
   end
@@ -19,5 +19,18 @@ describe "the pretty_location plugin" do
   it "prepends version names to generated location" do
     uploaded_file = @uploader.upload(fakeio(filename: "foo.jpg"), version: :thumb)
     assert_match %r{^thumb-[\w-]+\.jpg$}, uploaded_file.id
+  end
+
+  it "includes only the inner class in location by default" do
+    uploaded_file = @uploader.upload(fakeio, record: NameSpaced::OpenStruct.new(id: 123), name: :avatar)
+
+    assert_match %r{^openstruct/123/avatar/[\w-]+$}, uploaded_file.id
+  end
+
+  it "includes class namespace when :namespace is set" do
+    @uploader.class.plugin :pretty_location, namespace: "_"
+    uploaded_file = @uploader.upload(fakeio, record: NameSpaced::OpenStruct.new(id: 123), name: :avatar)
+
+    assert_match %r{^namespaced_openstruct/123/avatar/[\w-]+$}, uploaded_file.id
   end
 end